finished unit tests and appError

+++ src/modules/todo/useCases/deteleToDo/DeleteToDoUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 import { IToDosRepository } from '../../../../modules/todo/repositories/IToDoRepository';
+import AppError from '../../../../shared/errors/AppError';
 
 @injectable()
 class DeleteToDoUseCase {
@@ -10,6 +11,12 @@ class DeleteToDoUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const toDo = await this.todosRepository.findById(id);
+
+    if(!toDo) {
+      throw new AppError('ToDo not found');
+    }
+
     await this.todosRepository.delete(id);
   }
 }
